refactor(plans): estrai helper pure in plans-shared.ts per i client component
Problema: client component (export-csv-button, storico-client) che importavano
`canUsePro` o `type Plan` da @/lib/plans trascinavano l'intero modulo nel
client bundle. plans.ts importa @/db (drizzle + postgres), che non e'
risolvibile nel browser → build error "Module not found: postgres".

Soluzione:
- Nuovo `src/lib/plans-shared.ts` con TUTTI gli helper puri (Plan, canUsePro,
  canEmit, canAddCatalogItem, canUseApi, isDeveloperPlan, isTrialExpired,
  TRIAL_DAYS, STARTER_CATALOG_LIMIT, API_KEY_LIMITS, DEVELOPER_MONTHLY_LIMITS).
  Niente import server-side.
- `src/lib/plans.ts` re-esporta i pure per backward compat; mantiene `getPlan`
  e `assertProPlan` (server-only) che importano @/db.
- Aggiorno gli import nei 3 client component:
  - pro-feature-gate.tsx
  - export-csv-button.tsx
  - storico-client.tsx
  → tutti puntano a plans-shared.

I 14 importer server-side (route handler, server actions, RSC) restano
invariati: continuano a importare da @/lib/plans.
