+++ src/app/dashboard/storico/export-csv-button.tsx
@@ -1,7 +1,7 @@
 import Link from "next/link";
 import { Download } from "lucide-react";
 import { Button } from "@/components/ui/button";
-import { canUsePro, type Plan } from "@/lib/plans";
+import { canUsePro, type Plan } from "@/lib/plans-shared";
 
 interface ExportCsvButtonProps {
   readonly plan: Plan;
